Edit post

+++ src/components/RemoveModal.js
@@ -10,17 +10,21 @@ import Dialog, {
   DialogTitle,
 } from 'material-ui/Dialog'
 
-const RemoveComment = (props) => (
+const RemoveModal = (props) => (
   <Dialog
     open={props.open}
     onRequestClose={(e) => props.close(false)}
   >
     <DialogTitle>ATTENTION:</DialogTitle>
-      <DialogContent >Are you sure you want to remove this comment ?</DialogContent>
+    <DialogContent >{`Are you sure you want to remove this ${props.mode} ?`}</DialogContent>
       <DialogActions>
         <Button onClick={() => props.close(false)} color="primary">Cancel</Button>
         <Button onClick={ () => {
-            props.deleteComment(props.id)
+            if(props.mode === 'comment'){
+              props.deleteComment(props.id)
+            }else {
+              
+            }
             props.close(false)
           }
         }
@@ -29,15 +33,16 @@ const RemoveComment = (props) => (
   </Dialog>
 )
 
-RemoveComment.prototype = {
+RemoveModal.prototype = {
   close:PropTypes.func.isRequired,
   deleteComment:PropTypes.func.isRequired,
   open:PropTypes.bool.isRequired,
-  id:PropTypes.string.isRequired
+  id:PropTypes.string.isRequired,
+  mode:PropTypes.string.isRequired,
 }
  
 const mapDispatchToProps = (dispatch) => ({
   deleteComment: (id) => dispatch(removeComment(id)),
 })
 
-export default connect(null, mapDispatchToProps)(RemoveComment)
+export default connect(null, mapDispatchToProps)(RemoveModal)
